+++ src/application_settings/container_section_base.py
@@ -1,10 +1,8 @@
 """Base class for sections to be added to containers and container sections for configuration and settings."""
 import sys
 from abc import ABC, abstractmethod
-from dataclasses import replace
+from dataclasses import is_dataclass, replace
 from typing import Any, Literal, Optional, cast
-
-from pydantic.dataclasses import dataclass
 
 if sys.version_info >= (3, 11):
     from typing import Self
@@ -15,18 +13,17 @@ else:
 SectionTypeStr = Literal["Config", "Settings"]
 
 
-@dataclass(frozen=True)
 class ContainerSectionBase(ABC):
     """Base class for all ContainerSection classes"""
 
     @classmethod
     @abstractmethod
     def kind_string(cls) -> SectionTypeStr:
-        "Return either 'Config' or 'Settings'"
+        """Return either 'Config' or 'Settings'"""
 
     @classmethod
     def get(cls) -> Self:
-        """Get the singleton; if not existing, create it. Reading file only done for a container."""
+        """Get the singleton; if not existing, create it. Loading from file only done for a container."""
 
         if (_the_container_or_none := cls._get()) is None:
             # no config section has been made yet,
@@ -41,7 +38,13 @@ class ContainerSectionBase(ABC):
 
     @classmethod
     def update(cls, changes: dict[str, Any]) -> Self:
-        "Update the settings with data specified in changes; not meant for config"
+        """Update the settings with data specified in changes; not meant for config.
+
+        Raises:
+
+        * TypeError: if update is called upon a Config class
+        * RuntimeError: if filepath() == None
+        """
         return cls.get()._update(changes)._set()  # pylint: disable=protected-access
 
     @classmethod
@@ -63,10 +66,11 @@ class ContainerSectionBase(ABC):
         print(
             f"Section {cls.__name__} accessed before data has been set by the application."
         )
-        return cls().set({})
+        return cls.set({})
 
     def _set(self) -> Self:
         """Store the singleton."""
+        _check_dataclass_decorator(self)
         _ALL_CONTAINER_SECTION_SINGLETONS[id(self.__class__)] = self
         subsections = [
             attr
@@ -79,7 +83,22 @@ class ContainerSectionBase(ABC):
 
     def _update(self, changes: dict[str, Any]) -> Self:
         "Update parameters and sections with data specified in changes; not meant for config"
-        return replace(self, **changes)
+        # in self._set(), which normally is always executed, we ensured that
+        # self is a dataclass instance
+        return replace(self, **changes)  # type: ignore[type-var]
+
+
+def _check_dataclass_decorator(obj: Any) -> None:
+    if not (is_dataclass(obj)):
+        raise TypeError(
+            f"{obj} is not a dataclass instance; did you forget to add "
+            f"'@dataclass(frozen=True)' when you defined {obj.__class__}?."
+        )
+    if not obj.__class__.__dataclass_params__.frozen:
+        raise TypeError(
+            f"{obj} is not a frozen dataclass instance; did you forget "
+            f"to add '(frozen=True)' when you defined {obj.__class__}?."
+        )
 
 
 _ALL_CONTAINER_SECTION_SINGLETONS: dict[int, ContainerSectionBase] = {}