+++ src/application_settings/settings_base.py
@@ -1,8 +1,6 @@
 """Module for handling settings."""
 from typing import TypeVar
 
-from pydantic.dataclasses import dataclass
-
 from application_settings.container_base import ContainerBase, FileFormat
 from application_settings.container_section_base import (
     ContainerSectionBase,
@@ -13,26 +11,24 @@ SettingsT = TypeVar("SettingsT", bound="SettingsBase")
 SettingsT.__doc__ = "Represents SettingsBase and all subclasses"
 
 
-@dataclass(frozen=True)
 class SettingsSectionBase(ContainerSectionBase):
     """Base class for all SettingsSection classes (so that we can bound a TypeVar)"""
 
     @classmethod
     def kind_string(cls) -> SectionTypeStr:
-        "Return 'Settings'"
+        """Return 'Settings'"""
         return "Settings"
 
 
-@dataclass(frozen=True)
 class SettingsBase(ContainerBase):
     """Base class for main Settings class"""
 
     @classmethod
     def kind_string(cls) -> SectionTypeStr:
-        "Return 'Settings'"
+        """Return 'Settings'"""
         return "Settings"
 
     @classmethod
     def default_file_format(cls) -> FileFormat:
-        "Return the default file format"
+        """Return the default file format"""
         return FileFormat.JSON