+++ CHANGELOG.md
@@ -9,7 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Check for presence of the dataclass decorator
+  (Issue [#76](https://github.com/StockwatchDev/application_settings/issues/76))
+
 ### Changed
+
+- pydantic.dataclasses.dataclass and pydantic.ValidationError now exported from our
+  namespace (Issue [[#36](https://github.com/StockwatchDev/application_settings/issues/36)])
+- Exceptions added to docstrings (Issue 
+  [[#16](https://github.com/StockwatchDev/application_settings/issues/16)])
 
 ### Fixed